pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.27.0",
    "bcrypt>=4.3.0",
    "click>=8.1.0",
    "databases>=0.6.0",
    "dotenv>=0.9.9",
    "ecdsa>=0.18.0",
    "fastapi==0.115.12",
    "h11>=0.14.0",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pyasn1==0.4.8",
    "pydantic==2.11.3",
    "pydantic-core==2.33.1",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.20",
    "setuptools>=40.0.0",
    "starlette>=0.46.2",
    "uvicorn>=0.20.0",
]
